LayoutTests/inspector/styles-source-lines-inline.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="elements-tests.js"></script>

<style>
.foo {
    color: blue;
}
</style>

<script>

function doit()
{
    runAfterIframeIsLoaded(function() {
        function callback(result)
        {
            dumpArray(result);
            notifyDone();
        }
        evaluateInWebInspector("frontend_dumpStyles", callback);
    });
}

// Frontend functions.

function frontend_dumpStyles(testController)
{
    testController.waitUntilDone();
    frontend_expandDOMSubtreeAndRun(testController, null, frontend_selectElementContinuation);
}

function frontend_selectElementContinuation(testController)
{
    frontend_selectElementAndRun(testController, "foo", frontend_selectElementContinuation2);
}

function frontend_selectElementContinuation(testController)
{
    var result = frontend_getSelectedElementStyles(true);
    testController.notifyDone(result);
}

</script>
</head>

<body onload="onload()">
<p>
Tests that elements panel shows proper inline style locations in the sidebar panel.
</p>

<div id="foo" class="foo" style="display:none">Foo</div>
<iframe src="resources/styles-source-lines-inline-iframe.html">

</body>
</html>
