LayoutTests/inspector/styles-source-lines-inline-expected.txt

Tests that elements panel shows proper inline style locations in the sidebar panel.


[expanded] element.style { (element’s “style” attribute)
display: none;

======== Matched CSS Rules ========
[expanded] .foo { (styles-source-lines-inline.html:7)
color: blue;

[expanded] div { (user agent stylesheet)
/-- overloaded --/ display: block;


[expanded] element.style { (element’s “style” attribute)

======== Matched CSS Rules ========
[expanded] .bar { (resources/styles-source-lines-inline-iframe.html:4)
color: red;

[expanded] div { (user agent stylesheet)
display: block;
